+++ javascript/node/selenium-webdriver/lib/command.js
@@ -145,6 +145,7 @@ const Name = {
   IS_ELEMENT_DISPLAYED: 'isElementDisplayed',
   GET_ELEMENT_LOCATION: 'getElementLocation',
   GET_ELEMENT_LOCATION_IN_VIEW: 'getElementLocationOnceScrolledIntoView',
+  GET_ELEMENT_RECT: 'getElementRect',
   GET_ELEMENT_SIZE: 'getElementSize',
   GET_ELEMENT_ATTRIBUTE: 'getElementAttribute',
   GET_ELEMENT_VALUE_OF_CSS_PROPERTY: 'getElementValueOfCssProperty',